Revert "Always run "Previous Minor" job (#325)"
This reverts commit 7ac0af088330c30a7027d5ed6f63b93a38ab3ba7.
