+++ tests/Domain/Ci/Job/IntegratedTestOnLatestLtsCiJobTest.php
@@ -3,20 +3,15 @@
 namespace Acquia\Orca\Tests\Domain\Ci\Job;
 
 use Acquia\Orca\Domain\Ci\Job\AbstractCiJob;
-use Acquia\Orca\Domain\Ci\Job\Helper\RedundantJobChecker;
 use Acquia\Orca\Domain\Ci\Job\IntegratedTestOnLatestLtsCiJob;
 use Acquia\Orca\Domain\Composer\Version\DrupalCoreVersionResolver;
-use Acquia\Orca\Enum\CiJobEnum;
 use Acquia\Orca\Enum\DrupalCoreVersionEnum;
 use Acquia\Orca\Helper\EnvFacade;
 use Acquia\Orca\Helper\Process\ProcessRunner;
 use Acquia\Orca\Tests\Domain\Ci\Job\_Helper\CiJobTestBase;
-use Prophecy\Argument;
 use Symfony\Component\Console\Output\OutputInterface;
 
 /**
- * @property \Acquia\Orca\Domain\Ci\Job\Helper\RedundantJobChecker|\Prophecy\Prophecy\ObjectProphecy $redundantJobChecker
- *
  * @property \Acquia\Orca\Helper\EnvFacade|\Prophecy\Prophecy\ObjectProphecy $envFacade
  */
 class IntegratedTestOnLatestLtsCiJobTest extends CiJobTestBase {
@@ -26,10 +21,6 @@ class IntegratedTestOnLatestLtsCiJobTest extends CiJobTestBase {
     $this->envFacade = $this->prophesize(EnvFacade::class);
     $this->output = $this->prophesize(OutputInterface::class);
     $this->processRunner = $this->prophesize(ProcessRunner::class);
-    $this->redundantJobChecker = $this->prophesize(RedundantJobChecker::class);
-    $this->redundantJobChecker
-      ->isRedundant(Argument::any())
-      ->willReturn(FALSE);
     parent::setUp();
   }
 
@@ -38,8 +29,7 @@ class IntegratedTestOnLatestLtsCiJobTest extends CiJobTestBase {
     $env_facade = $this->envFacade->reveal();
     $output = $this->output->reveal();
     $process_runner = $this->processRunner->reveal();
-    $redundant_job_checker = $this->redundantJobChecker->reveal();
-    return new IntegratedTestOnLatestLtsCiJob($drupal_core_version_resolver, $env_facade, $output, $process_runner, $redundant_job_checker);
+    return new IntegratedTestOnLatestLtsCiJob($drupal_core_version_resolver, $env_facade, $output, $process_runner);
   }
 
   public function testBasicConfiguration(): void {
@@ -61,22 +51,6 @@ class IntegratedTestOnLatestLtsCiJobTest extends CiJobTestBase {
     $job = $this->createJob();
 
     $this->runInstallPhase($job);
-  }
-
-  public function testRedundantJob(): void {
-    $this->redundantJobChecker
-      ->isRedundant(CiJobEnum::INTEGRATED_TEST_ON_LATEST_LTS())
-      ->willReturn(TRUE);
-    $this->output
-      ->writeln(Argument::any())
-      ->shouldBeCalledTimes(2);
-    $this->processRunner
-      ->runOrca(Argument::any())
-      ->shouldNotBeCalled();
-    $job = $this->createJob();
-
-    $this->runScriptPhase($job);
-    $this->runScriptPhase($job);
   }
 
   public function testInstallOverrideProfile(): void {
